requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg==3.1.13
python-dotenv==1.0.0
celery==5.3.6
redis==5.0.1
Pillow==10.1.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
shortuuid==1.0.11
Faker==20.1.0
